Decompose Leaderboard into LeaderboardForm
